[app-server] refactor: split API types into v1 and v2
